Split ROM specs on v3 and v4
