+++ Gemfile
@@ -10,7 +10,7 @@ group :development, :test do
   gem 'dm-sqlite-adapter'
   gem 'dry-struct'
   gem 'dry-types'
-  gem 'rom', '~> 3.0'
+  gem 'rom', '~> 4.0'
   gem 'rom-repository'
   gem 'rom-sql'
   gem 'sequel'
